requirements.txt
django
mysqlclient
pillow
django-ninja
django-cors-headers
requests
orjson

django-redis
django-celery-results
django-celery-beat

celery==5.2.3
redis
gunicorn
uwsgi

django-ninja-jwt
python-jose[cryptography]

python-dotenv
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0